backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
together>=0.2.8
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
cinemagoer==2023.5.1
tqdm==4.66.1




